Export and import class in index.js
